corrections/build bug fixes, failing tests skiipped, now builds corrrectly for PHP 7.x, and passing phpunit tests under 8.x
