+++ .gitignore
@@ -2,3 +2,4 @@ vendor
 composer.lock
 .DS_Store
 build/
+*.cache
